apps/user/api_endpoints/User/UserUpdatePasswordLink/serializer.py
from rest_framework.serializers import ModelSerializer

from apps.user.models import User


class UserUpdatePasswordLinkSerializer(ModelSerializer):

    class Meta:
        model = User
        fields = ['email']
